// src/lib/types/YieldHistory.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { YieldDataPoint } from "./YieldDataPoint";

export type YieldHistory = { player_id: number, player_name: string, nation: string | null, yield_type: string, data: Array<YieldDataPoint>, };